Create annotated exception class
