+++ validators/structure_validator.pyi
@@ -2,10 +2,6 @@ from dodona.translator import Translator
 
 
 class NotTheSame(Exception):
-    msg: str
-    line: int
-    trans: Translator
-
     def message_str(self) -> str: ...
 
     def annotation_str(self) -> str: ...
